chatbot_with_pretrained_model/requirement.txt
PyPDF2
sentence-transformers
faiss-cpu